+++ pystache/tests/common.py
@@ -7,10 +7,9 @@ Provides test-related code that can be used by all tests.
 
 import os
 
-import examples
 import pystache
 from pystache import defaults
-
+from pystache.tests import examples
 
 # Save a reference to the original function to avoid recursion.
 _DEFAULT_TAG_ESCAPE = defaults.TAG_ESCAPE
@@ -18,9 +17,14 @@ _TESTS_DIR = os.path.dirname(pystache.tests.__file__)
 
 DATA_DIR = os.path.join(_TESTS_DIR, 'data')  # i.e. 'pystache/tests/data'.
 EXAMPLES_DIR = os.path.dirname(examples.__file__)
-SOURCE_DIR = os.path.dirname(pystache.__file__)
-PROJECT_DIR = os.path.join(SOURCE_DIR, '..')
+PACKAGE_DIR = os.path.dirname(pystache.__file__)
+PROJECT_DIR = os.path.join(PACKAGE_DIR, '..')
 SPEC_TEST_DIR = os.path.join(PROJECT_DIR, 'ext', 'spec', 'specs')
+# TEXT_DOCTEST_PATHS: the paths to text files (i.e. non-module files)
+# containing doctests.  The paths should be relative to the project directory.
+TEXT_DOCTEST_PATHS = ['README.rst']
+
+UNITTEST_FILE_PREFIX = "test_"
 
 
 def html_escape(u):
@@ -41,6 +45,80 @@ def html_escape(u):
 
 def get_data_path(file_name):
     return os.path.join(DATA_DIR, file_name)
+
+
+# Functions related to get_module_names().
+
+def _find_files(root_dir, should_include):
+    """
+    Return a list of paths to all modules below the given directory.
+
+    Arguments:
+
+      should_include: a function that accepts a file path and returns True or False.
+
+    """
+    paths = []  # Return value.
+
+    is_module = lambda path: path.endswith(".py")
+
+    # os.walk() is new in Python 2.3
+    #   http://docs.python.org/library/os.html#os.walk
+    for dir_path, dir_names, file_names in os.walk(root_dir):
+        new_paths = [os.path.join(dir_path, file_name) for file_name in file_names]
+        new_paths = filter(is_module, new_paths)
+        new_paths = filter(should_include, new_paths)
+        paths.extend(new_paths)
+
+    return paths
+
+
+def _make_module_names(package_dir, paths):
+    """
+    Return a list of fully-qualified module names given a list of module paths.
+
+    """
+    package_dir = os.path.abspath(package_dir)
+    package_name = os.path.split(package_dir)[1]
+
+    prefix_length = len(package_dir)
+
+    module_names = []
+    for path in paths:
+        path = os.path.abspath(path)  # for example <path_to_package>/subpackage/module.py
+        rel_path = path[prefix_length:]  # for example /subpackage/module.py
+        rel_path = os.path.splitext(rel_path)[0]  # for example /subpackage/module
+
+        parts = []
+        while True:
+            (rel_path, tail) = os.path.split(rel_path)
+            if not tail:
+                break
+            parts.insert(0, tail)
+        # We now have, for example, ['subpackage', 'module'].
+        parts.insert(0, package_name)
+        module = ".".join(parts)
+        module_names.append(module)
+
+    return module_names
+
+
+def get_module_names(package_dir=None, should_include=None):
+    """
+    Return a list of fully-qualified module names in the given package.
+
+    """
+    if package_dir is None:
+        package_dir = PACKAGE_DIR
+
+    if should_include is None:
+        should_include = lambda path: True
+
+    paths = _find_files(package_dir, should_include)
+    names = _make_module_names(package_dir, paths)
+    names.sort()
+
+    return names
 
 
 class AssertStringMixin: