+++ pystache/tests/data/views.py
@@ -1,5 +1,10 @@
 # coding: utf-8
 
+"""
+TODO: add a docstring.
+
+"""
+
 from pystache import TemplateSpec
 
 class SayHello(object):