+++ pystache/tests/examples/complex.py
@@ -1,3 +1,9 @@
+
+"""
+TODO: add a docstring.
+
+"""
+
 class Complex(object):
 
     def header(self):