+++ pystache/tests/examples/delimiters.py
@@ -1,3 +1,9 @@
+
+"""
+TODO: add a docstring.
+
+"""
+
 class Delimiters(object):
 
     def first(self):