+++ pystache/tests/examples/escaped.py
@@ -1,3 +1,9 @@
+
+"""
+TODO: add a docstring.
+
+"""
+
 class Escaped(object):
 
     def title(self):