+++ pystache/tests/examples/unicode_output.py
@@ -1,5 +1,10 @@
 # encoding: utf-8
 
+"""
+TODO: add a docstring.
+
+"""
+
 class UnicodeOutput(object):
 
     def name(self):